implicit conversations

+++ Saxon.C.API/SaxonCGlue.c
@@ -173,7 +173,7 @@ void setParameter(sxnc_parameter **parameters, int *parLen, int *parCap,
     }
     *parameters = temp;
   }
-  int nameLen = strlen(name) + 7;
+  int nameLen = (int) strlen(name) + 7;
   char *newName = malloc(sizeof(char) * nameLen);
   snprintf(newName, nameLen, "%s%s", "param:", name);
   (*parameters)[(*parLen) - 1].name = (char *)newName;
@@ -199,7 +199,7 @@ void setProperty(sxnc_property **properties, int *propLen, int *propCap,
     }
     *properties = temp;
   }
-  int nameLen = strlen(name) + 1;
+  int nameLen = (int) strlen(name) + 1;
   char *newName = (char *)malloc(sizeof(char) * nameLen);
   snprintf(newName, nameLen, "%s", name);
   char *newValue = (char *)malloc(sizeof(char) * strlen(value) + 1);
