+++ Saxon.C.API/SaxonProcessor.cpp
@@ -368,7 +368,7 @@ bool SaxonProcessor::isSchemaAwareProcessor() {
 void SaxonProcessor::applyConfigurationProperties() {
   SaxonProcessor::attachCurrentThread();
   if (configProperties.size() > 0) {
-    int size = configProperties.size();
+    int size = (int) configProperties.size();
     int64_t processorDataRef = createProcessorDataWithCapacity(
         SaxonProcessor::sxn_environ->thread, size);
 
@@ -510,7 +510,7 @@ int64_t SaxonProcessor::createParameterJArray(
     std::map<std::string, XdmValue *> parameters,
     std::map<std::string, std::string> properties, int additions) {
 
-  int size = parameters.size() + properties.size() + additions;
+  int size = (int) parameters.size() + (int) properties.size() + additions;
 
   if (getenv("SAXONC_DEBUG_FLAG")) {
     std::cerr << "Properties size: " << properties.size() << std::endl;
@@ -576,7 +576,7 @@ int64_t SaxonProcessor::createParameterJArray(
 int64_t SaxonProcessor::createParameterJArray2(
     std::map<std::string, XdmValue *> parameters) {
 
-  int size = parameters.size();
+  int size = (int) parameters.size();
   int64_t processorDataRef = createProcessorDataWithCapacity(
       SaxonProcessor::sxn_environ->thread, size);
 
@@ -1183,7 +1183,7 @@ SaxonProcessor::makeMap(std::map<XdmAtomicValue *, XdmValue *> dataMap) {
   int64_t mapDataRef = SXN_UNSET;
 
   mapDataRef =
-      j_create_mapDataWithCapacity(sxn_environ->thread, dataMap.size());
+      j_create_mapDataWithCapacity(sxn_environ->thread, (int) dataMap.size());
 
   for (std::map<XdmAtomicValue *, XdmValue *>::iterator iter = dataMap.begin();
        iter != dataMap.end(); ++iter) {
@@ -1217,7 +1217,7 @@ XdmMap *SaxonProcessor::makeMap2(std::map<std::string, XdmValue *> dataMap) {
   int64_t mapDataRef = SXN_UNSET;
 
   mapDataRef =
-      j_create_mapDataWithCapacity(sxn_environ->thread, dataMap.size());
+      j_create_mapDataWithCapacity(sxn_environ->thread, (int) dataMap.size());
 
   for (std::map<std::string, XdmValue *>::iterator iter = dataMap.begin();
        iter != dataMap.end(); ++iter) {
